Remove redundant dependency on express in ./lib/device/js
express was being required in lib/device.js but not actually used. Since it is also declared as a prod dependency it will be installed for all implementations and so will potentially unnecessarily bloat the dependency tree.

Moving the dependency to a dev dependency allows the files still to be tested in an express context for development, but will not result in an additional and potentially redundant installation for production implementations.